Move user spec helpers from plugins
